Split the script in smaller functions. [ci skip]

+++ build.sh
@@ -1,32 +1,63 @@
-#! /bin/sh -e
+#! /bin/sh 
 
-FS_DIR=filesystem
-ISO_DIR=iso_image
-CONFIG_DIR=config
+set -e
+
+FS_DIR=root
+ISO_DIR=image
 IMAGE_NAME=nitrux
 
-# Prepare the root filesystem.
 
-mkdir -p filesystem
+# Function for running commands in a chroot.
 
-wget -q http://cdimage.ubuntu.com/ubuntu-base/releases/18.04/release/ubuntu-base-18.04-base-amd64.tar.gz
-tar xf *.tar.gz -C filesystem
+run_chroot() {
+
+	clean () {
+		for d in $FS_DIR/*; do
+
+			mountpoint -q $d && \
+				umount -Rf $d
+
+		done
+	}
+
+	trap clean EXIT HUP INT TERM
+
+	mount -t proc . $FS_DIR/proc
+	mount -t sysfs . $FS_DIR/sys
+	mount -t devtmpfs . $FS_DIR/dev
+	mkdir -p $FS_DIR/dev/pts
+	mount -t devpts . $FS_DIR/dev/pts
+
+	if [ -f $1 && -x $1 ]; then
+		cp $1 $FS_DIR/
+		chroot $FS_DIR/ /$1
+		rm -r $FS_DIR/config.sh
+	else
+		chroot $FS_DIR/ $1
+	fi
+
+	clean
+
+}
 
 
-# Run a command in a chroot safely.
+# Prepare the directory were the filesystem will be created.
+
+mkdir -p $FS_DIR
+
+wget -O base.tar.gz -q http://cdimage.ubuntu.com/ubuntu-base/releases/18.04/release/ubuntu-base-18.04-base-amd64.tar.gz
+tar xf base.tar.gz -C $FS_DIR
+
 rm -rf $FS_DIR/dev/*
-cp /etc/resolv.conf filesystem/etc
+cp /etc/resolv.conf $FS_DIR/etc
 
-mount -t proc none $FS_DIR/proc || exit 1
-mount -t devtmpfs none $FS_DIR/dev || exit 1
-mkdir -p /dev/pts
-mount -t devpts none $FS_DIR/dev/pts || exit 1
 
-cp $CONFIG_DIR/config.sh $FS_DIR
-chroot $FS_DIR/ sh -c /config.sh
-rm -r $FS_DIR/config.sh
+# Create the filesystem.
 
-umount -f $FS_DIR/dev/pts $FS_DIR/dev $FS_DIR/proc
+run_chroot ./bootstrap.sh
+
+
+# Copy the initramfs and the kernel to $ISO_DIR.
 
 cp $FS_DIR/vmlinuz $ISO_DIR/boot/kernel
 cp $FS_DIR/initrd.img $ISO_DIR/boot/initramfs
@@ -45,6 +76,7 @@ rm -rf $FS_DIR/tmp/* \
 # Compress the root filesystem.
 
 (sleep 300; echo +) &
+
 echo "Compressing the root filesystem"
 mkdir -p $ISO_DIR/casper
 mksquashfs $FS_DIR $ISO_DIR/casper/filesystem.squashfs -comp xz -no-progress
@@ -53,27 +85,27 @@ kill $! || true
 
 # Create the ISO image.
 
-cd $ISO_DIR
-echo -n $(du -sx --block-size=1 . | tail -n 1 | awk '{ print $1 }') > casper/filesystem.size
+(
+	cd $ISO_DIR
+	echo -n $(du -sx --block-size=1 . | tail -n 1 | awk '{ print $1 }') > casper/filesystem.size
 
-xorriso -as mkisofs -r -J -l \
-	-V '#NITRUX' \
-	-isohybrid-mbr boot/isolinux/isohdpfx.bin \
-	-c boot/isolinux/boot.cat \
-	-b boot/isolinux/isolinux.bin \
-	-no-emul-boot \
-	-boot-load-size 4 \
-	-boot-info-table \
-	-eltorito-alt-boot \
-	-e boot/grub/efi.img \
-	-no-emul-boot \
-	-isohybrid-gpt-basdat \
-	-o ../$IMAGE_NAME.iso .
-
-cd ..
+	xorriso -as mkisofs -r -J -l \
+		-V 'NITRUX_LIVE' \
+		-isohybrid-mbr boot/isolinux/isohdpfx.bin \
+		-c boot/isolinux/boot.cat \
+		-b boot/isolinux/isolinux.bin \
+		-no-emul-boot \
+		-boot-load-size 4 \
+		-boot-info-table \
+		-eltorito-alt-boot \
+		-e boot/grub/efi.img \
+		-no-emul-boot \
+		-isohybrid-gpt-basdat \
+		-o ../$IMAGE_NAME.iso .
+)
 
 zsyncmake $IMAGE_NAME.iso
-echo "zsync|http://server.domain/path/your.iso.zsync" | dd of=$IMAGE_NAME.iso bs=1 seek=33651 count=512 conv=notrunc
+echo "http://server.domain/path/your.iso.zsync" | dd of=$IMAGE_NAME.iso bs=1 seek=33651 count=512 conv=notrunc
 
 sha256sum $IMAGE_NAME.iso > checksum
 curl -i -F filedata=@checksum -F filedata=@$IMAGE_NAME.iso https://transfer.sh | sed 's/http/\nhttp/g' | grep http > urls
